+++ js/index.js
@@ -15,22 +15,6 @@
 })();
 
 /**
- * Open modal
- */
-(function () {
-  const attribute = 'data-js-modal';
-  const modalTriggers = document.querySelectorAll(`[${attribute}]`);
-  modalTriggers.forEach((modalTrigger) => {
-    const modal = document.getElementById(modalTrigger.getAttribute(attribute));
-    if (modal) {
-      modalTrigger.addEventListener('click', (event) => {
-        modal.style.display = 'block';
-      });
-    }
-  });
-})();
-
-/**
  * Cancel single item
  */
 (function () {
