+++ js/modal.js
@@ -1,0 +1,14 @@
+const modal = () => {
+  const attribute = 'data-js-modal';
+  const modalTriggers = document.querySelectorAll(`[${attribute}]`);
+  modalTriggers.forEach((modalTrigger) => {
+    const modal = document.getElementById(modalTrigger.getAttribute(attribute));
+    if (modal) {
+      modalTrigger.addEventListener('click', (event) => {
+        modal.style.display = 'block';
+      });
+    }
+  });
+};
+
+export default modal;
